refactor(email): Refactor the email sending
Remove SMTP connection pool in favor of direct connections

BREAKING CHANGE: Replace SMTP connection pool with direct connections

The SMTP connection pool has been removed in favor of creating new
connections for each email send operation. This change:

- Removes Pool struct and related code
- Adds new Mailer struct for handling email operations
- Updates all email sending functions to use new Mailer
- Removes pool size configuration
- Simplifies connection management

Rationale:
- SMTP servers often don't handle long-lived connections well
- Connection pool adds complexity without significant benefits
- Direct connections are more reliable for email sending
- Email sending is typically not high-frequency enough to warrant pooling

Migration:
- Replace getSmtpPool() calls with getMailer()
- Update any direct Pool usage to use Mailer instead
